Add endpoint for creating and getting a list of users

+++ homecontrol_api/database/users.py
@@ -19,10 +19,10 @@ class UsersDBConnection(DatabaseConnection):
         self._session.commit()
         self._session.refresh(user)
         return user
-    
+
     def get(self, user_id: str) -> UserInDB:
         """Returns UserInDB given a user's ID
-        
+
         Args:
             user_id (str): ID of the user
 
@@ -39,10 +39,10 @@ class UsersDBConnection(DatabaseConnection):
         if not user:
             raise DatabaseEntryNotFoundError(f"User with id '{user_id}' was not found")
         return user
-    
+
     def get_by_username(self, username: str) -> UserInDB:
         """Returns UserInDB given a user's username
-        
+
         Args:
             username (str): Username of the user
 
@@ -50,19 +50,25 @@ class UsersDBConnection(DatabaseConnection):
             UserInDB: Info about the user
 
         Raises:
-            DatabaseEntryNotFoundError: If the user isn't found    
+            DatabaseEntryNotFoundError: If the user isn't found
         """
         user = (
-            self._session.query(UserInDB).filter(UserInDB.username ==  username).first()
+            self._session.query(UserInDB).filter(UserInDB.username == username).first()
         )
         if not user:
-            raise DatabaseEntryNotFoundError(f"User with username '{username}' was not found")
+            raise DatabaseEntryNotFoundError(
+                f"User with username '{username}' was not found"
+            )
         return user
-    
+
     def get_all(self) -> list[UserInDB]:
         """Returns a list of information about all users"""
         return self._session.query(UserInDB).all()
-    
+
+    def count(self) -> int:
+        """Returns the number of users in the database"""
+        return self._session.query(UserInDB).count()
+
     def delete(self, user_id: str):
         """Deletes a UserInDB given the users's id
 
@@ -73,15 +79,10 @@ class UsersDBConnection(DatabaseConnection):
             DatabaseEntryNotFoundError: If the user isn't found
         """
         rows_deleted = (
-            self._session.query(UserInDB)
-            .filter(UserInDB.id == UUID(user_id))
-            .delete()
+            self._session.query(UserInDB).filter(UserInDB.id == UUID(user_id)).delete()
         )
 
         if rows_deleted == 0:
-            raise DatabaseEntryNotFoundError(
-                f"User with id '{user_id}' was not found"
-            )
+            raise DatabaseEntryNotFoundError(f"User with id '{user_id}' was not found")
 
         self._session.commit()
-
